fix: move var.tf to terraform.tfvars.example to allow project to run
In the state it was in terraform was trying to parse this file and
execute the statements in it, however as these were just variable
declarations, terraform was erroring out.

Thus this file is now moved to a .example extension which will cause it
not to be parsed as terraform _or_ as variables.  If this file is copied
to a variables.tf file it will be used as the source of truth for
variables.